flesh out the mocks module
